Extract tests to separate files 📂
